==> packages/wmb/wmb-0.1.24.tar.gz/wmb-0.1.24/analysis/integration/merfish_mc/00.prepare.ipynb <==
import pandas as pd
import numpy as np
from wmb import cemba
from ALLCools.plot import *

import pathlib

mc_annot = cemba.get_mc_annot()

def save_index(cells, file_name, downsample=100000, random_state=0):
    cells = pd.Series(cells)
    if cells.size > downsample:
        cells = cells.sample(downsample, random_state=random_state)
    cells.to_csv(file_name, index=None, header=False)


save_index(mc_annot.get_index('cell'),
           file_name='./mc_cells.txt',
           downsample=100000,
           random_state=0)

import joblib
import pathlib
import subprocess

mc_groups = pd.read_csv('../L1/Neuron/mc_integration_group.csv.gz',
                        index_col=0).squeeze()
rna_groups = pd.read_csv('../L1/Neuron/rna_integration_group.csv.gz',
                         index_col=0).squeeze()

mc_groups.value_counts()

for integration_group in rna_groups.unique():
    mc_cells = mc_groups[mc_groups == integration_group].index
    rna_cells = rna_groups[rna_groups == integration_group].index

    integration_group = f'InteGroup{integration_group}'
    out_dir = pathlib.Path(f'../L2/{integration_group}')
    out_dir.mkdir(exist_ok=True, parents=True)

    save_index(mc_cells,
               f'../L2/{integration_group}/mc_cells.txt',
               downsample=1000000,
               random_state=0)
    save_index(rna_cells,
               f'../L2/{integration_group}/rna_cells.txt',
               downsample=1000000,
               random_state=0)

cwd = pathlib.Path().absolute()
with open('Snakefile') as f, open('../L2/Snakefile', 'w') as out_f:
    snakefile_tmp = f.read()
    snakefile_tmp = snakefile_tmp.replace('REPLACE_TEMPLATE_DIR', str(cwd))
    snakefile_tmp = snakefile_tmp.replace('REPLACE_DATASET', 'AIBS_SMART')
    snakefile_tmp = snakefile_tmp.replace('REPLACE_CLUSTER_COL', str('L2'))
    snakefile_tmp = snakefile_tmp.replace('REPLACE_INTEGRATION_GROUP_KEY',
                                          str('L3'))
    snakefile_tmp = snakefile_tmp.replace('REPLACE_INTEGRATION_PLOT_KEY',
                                          str('L2'))
    snakefile_tmp = snakefile_tmp.replace(
        'REPLACE_CATEGORICAL_KEY', str(['L2', 'L3', 'DissectionRegion']))
    out_f.write(snakefile_tmp)

import joblib
import pathlib
import subprocess

L2_group = []
for path in pathlib.Path('../L2').glob('InteGroup*'):
    L2_group.append(str(path).split('/')[-1])

for l2_group in L2_group:
    mc_groups = pd.read_csv(f'../L2/{l2_group}/mc_integration_group.csv.gz',
                        index_col=0).squeeze()
    rna_groups = pd.read_csv(f'../L2/{l2_group}/rna_integration_group.csv.gz',
                         index_col=0).squeeze()
    
    for integration_group in rna_groups.unique():
        mc_cells = mc_groups[mc_groups == integration_group].index
        rna_cells = rna_groups[rna_groups == integration_group].index

        integration_group = f'{l2_group}_{integration_group}'
        if mc_cells.size <= 150 or rna_cells.size <= 150:
            print(integration_group)
        else:
            out_dir = pathlib.Path(f'../L3/{integration_group}')
            out_dir.mkdir(exist_ok=True, parents=True)

            save_index(mc_cells,
                       f'../L3/{integration_group}/mc_cells.txt',
                       downsample=1000000,
                       random_state=0)
            save_index(rna_cells,
                       f'../L3/{integration_group}/rna_cells.txt',
                       downsample=1000000,
                       random_state=0)


cwd = pathlib.Path().absolute()
with open('Snakefile') as f, open('../L3/Snakefile', 'w') as out_f:
    snakefile_tmp = f.read()
    snakefile_tmp = snakefile_tmp.replace('REPLACE_TEMPLATE_DIR', str(cwd))
    snakefile_tmp = snakefile_tmp.replace('REPLACE_DATASET', 'AIBS_SMART')
    snakefile_tmp = snakefile_tmp.replace('REPLACE_CLUSTER_COL', str('L3'))
    snakefile_tmp = snakefile_tmp.replace('REPLACE_INTEGRATION_GROUP_KEY',
                                          str('L3'))
    snakefile_tmp = snakefile_tmp.replace('REPLACE_INTEGRATION_PLOT_KEY',
                                          str('L3'))
    snakefile_tmp = snakefile_tmp.replace(
        'REPLACE_CATEGORICAL_KEY', str(['L3', 'DissectionRegion']))
    out_f.write(snakefile_tmp)

L2_group = []
for path in pathlib.Path('../L2').glob('InteGroup*'):
    L2_group.append(str(path).split('/')[-1])

for l2_group in L2_group:
    mc_groups = pd.read_csv(f'../L2/{l2_group}/mc_integration_group.csv.gz',
                        index_col=0).squeeze()
    rna_groups = pd.read_csv(f'../L2/{l2_group}/rna_integration_group.csv.gz',
                         index_col=0).squeeze()
    
    for integration_group in rna_groups.unique():
        mc_cells = mc_groups[mc_groups == integration_group].index
        rna_cells = rna_groups[rna_groups == integration_group].index

        integration_group = f'{l2_group}_{integration_group}'
        if mc_cells.size <= 150 or rna_cells.size <= 150:
            print(integration_group)
        else:
            out_dir = pathlib.Path(f'../L4/{integration_group}')
            out_dir.mkdir(exist_ok=True, parents=True)

            save_index(mc_cells,
                       f'../L4/{integration_group}/mc_cells.txt',
                       downsample=1000000,
                       random_state=0)
            save_index(rna_cells,
                       f'../L4/{integration_group}/rna_cells.txt',
                       downsample=1000000,
                       random_state=0)


cwd = pathlib.Path().absolute()
with open('Snakefile') as f, open('../L4/Snakefile', 'w') as out_f:
    snakefile_tmp = f.read()
    snakefile_tmp = snakefile_tmp.replace('REPLACE_TEMPLATE_DIR', str(cwd))
    snakefile_tmp = snakefile_tmp.replace('REPLACE_DATASET', 'AIBS_SMART')
    snakefile_tmp = snakefile_tmp.replace('REPLACE_CLUSTER_COL', str('L3'))
    snakefile_tmp = snakefile_tmp.replace('REPLACE_INTEGRATION_GROUP_KEY',
                                          str('L4'))
    snakefile_tmp = snakefile_tmp.replace('REPLACE_INTEGRATION_PLOT_KEY',
                                          str('L4'))
    snakefile_tmp = snakefile_tmp.replace(
        'REPLACE_CATEGORICAL_KEY', str(['L3', 'L4','DissectionRegion']))
    out_f.write(snakefile_tmp)
