==> packages/wmb/wmb-0.1.24.tar.gz/wmb-0.1.24/analysis/integration/merfish_mc/05.plot.ipynb <==
from ALLCools.clustering import *
from ALLCools.mcds import MCDS
from wmb import brain, aibs, broad, mm10

import pandas as pd
import anndata
import numpy as np

import matplotlib.pyplot as plt
from ALLCools.plot import *
import wmb
import scanpy as sc
from skimage.transform import rotate, SimilarityTransform, matrix_transform

ref_region_type = 'CEMBA'

adata_merge = anndata.read_h5ad('./adata/final_with_coords.h5ad')

## plot annot

fig, axes = plt.subplots(nrows=2,ncols=2,figsize=(10, 10),dpi=300)

merfish_adata = adata_merge[adata_merge.obs['Modality'] == 'merfish']
mC_adata = adata_merge[adata_merge.obs['Modality'] == 'mC']

ax = axes[0,0]
categorical_scatter(data=merfish_adata,
                        coord_base='tsne',
                        max_points=None,
                        hue=None,
                        scatter_kws=dict(color='lightgrey'),
                        ax=ax)

_ = categorical_scatter(
    data=mC_adata,
    ax=ax,
    coord_base='tsne',
    hue='L1_annot',
    text_anno='L1_annot',
    palette='tab20',
    max_points=None,)

ax.set(title = 'mC annot')

    
    
ax = axes[0,1]
categorical_scatter(data=mC_adata,
                        coord_base='tsne',
                        max_points=None,
                        hue=None,
                        scatter_kws=dict(color='lightgrey'),
                        ax=ax)

_ = categorical_scatter(
    data=merfish_adata,
    ax=ax,
    coord_base='tsne',
    hue='L1_annot_transfer',
    text_anno='L1_annot_transfer',
    palette='tab20',
    max_points=None,
)
ax.set(title = 'merfish transferred annot')

ax = axes[1,0]
categorical_scatter(data=merfish_adata,
                        coord_base='tsne',
                        max_points=None,
                        hue=None,
                        scatter_kws=dict(color='lightgrey'),
                        ax=ax)

_ = categorical_scatter(
    data=mC_adata,
    ax=ax,
    coord_base='tsne',
    hue='DissectionRegion',
    #text_anno='DissectionRegion',
    palette=wmb.brain.get_dissection_region_palette(region_type=ref_region_type),
    max_points=None,)

ax.set(title = 'mC L1 DissectionRegion')

    
    
ax = axes[1,1]
categorical_scatter(data=mC_adata,
                        coord_base='tsne',
                        max_points=None,
                        hue=None,
                        scatter_kws=dict(color='lightgrey'),
                        ax=ax)

_ = categorical_scatter(
    data=merfish_adata,
    ax=ax,
    coord_base='tsne',
    hue='DissectionRegion_transfer',
    #text_anno='DissectionRegion_transfer',
    palette=wmb.brain.get_dissection_region_palette(region_type=ref_region_type),
    max_points=None,
)
ax.set(title = 'merfish L1 transferred DissectionRegion')


fig.savefig('./fig/integration_annot.png')

## plot gene marker

gene_matrix = pd.read_pickle('mC_gene_matrix.pickle')
gene_matrix = gene_matrix.loc[:,adata_merge.var_names]

gene_matrix.head()

fig, ax = plt.subplots(nrows=1,ncols=1,figsize=(5, 5),dpi=300)

merfish_adata = adata_merge[adata_merge.obs['Modality'] == 'merfish']
mC_adata = adata_merge[adata_merge.obs['Modality'] == 'mC']


continuous_scatter(data=merfish_adata,
                        coord_base='tsne',
                        max_points=None,
                        hue=None,
                        scatter_kws=dict(color='lightgrey'),
                        ax=ax)

_ = continuous_scatter(
    data=mC_adata,
    ax=ax,
    coord_base='tsne',
    hue=gene_matrix['Adarb2'],
    text_anno='L1_annot',
    max_points=None,)


## plot spatial

adata = anndata.read_h5ad('./adata/merfish_input.h5ad')

adata.obs['annot'] = adata_merge.obs['L1_annot_transfer']

adata.obsm['X_spatial'] = adata.obsm['spatial'].copy()
del adata.obsm['spatial']

# rotate spatial
trans_mat = SimilarityTransform(rotation=15.5)
rotate_coord = matrix_transform(adata.obsm['X_spatial'], np.array(trans_mat))
adata.obsm['X_spatial_rotated'] = rotate_coord.copy()

adata

fig, ax = plt.subplots(figsize = (7,3), ncols = 1, dpi = 300, constrained_layout = True)


categorical_scatter(data=adata,
                        ax=ax,
                        coord_base='spatial_rotated',
                        hue='annot',
                    text_anno=None,
                    show_legend = True,
                       max_points = None)

ax.set(title = 'annot')

fig.savefig('./fig/spatial_annot.png')

all_cell_types = adata.obs['annot'].unique().to_list()
n_plots = len(all_cell_types)

fig, axes = plt.subplots(figsize=(24, 12),
                         nrows=6,
                         ncols=6,
                         dpi=300,
                         constrained_layout=True,
                         sharex=False)

for ax, ct in zip(axes.ravel(), all_cell_types):
    categorical_scatter(data=adata,
                            ax=ax,
                            coord_base='spatial_rotated',
                            hue=adata.obs['annot'].isin([ct]),
                            #text_anno='leiden',
                            show_legend=False,
                            palette = {
                                        True: 'red',
                                        False: 'lightgrey'
                                    },
                           max_points = None)
    ax.set( title = ct) 
    
for ax in axes.ravel()[n_plots:]:
    ax.axis('off')
    
fig.savefig('./fig/spatial_annot_seperate.png')
